* more quelling of gcc's bitching (don't we love -Wall)
git-svn-id: https://ascentemubackup.svn.sourceforge.net/svnroot/ascentemubackup/trunk@4369 8ad44209-8f18-4ceb-bdfa-99d0be514da2

+++ src/ascent-world/SpellEffects.cpp
@@ -4970,7 +4970,7 @@ void Spell::SpellEffectDispelMechanic(uint32 i)
 	{
 		if( unitTarget->m_auras[x] && !unitTarget->m_auras[x]->IsPositive())
 		{
-			if( unitTarget->m_auras[x]->GetSpellProto()->DispelType == sMisc || unitTarget->m_auras[x]->GetSpellProto()->EffectMechanic[x] == sMisc )
+			if( unitTarget->m_auras[x]->GetSpellProto()->DispelType == sMisc || unitTarget->m_auras[x]->GetSpellProto()->EffectMechanic[x] == (int32)sMisc )
 				unitTarget->m_auras[x]->Remove();
 		}
 	}
